+++ CHANGELOG.md
@@ -5,6 +5,7 @@ ADDED:
 - Ability to persist entities via `Worm::save()` and `Worm::delete()`
 - Ability to define if a relationship has `ON DELETE CASCADE` constraint
 - `IdentityMap::createObjectFromId` method
+- `SelectQueryBuilder::fetchColumn()` method
 
 CHANGED: